functional use of Alexa dialog functionality
