+++ app/lib/alexa/request_handler.rb
@@ -2,15 +2,20 @@ module Alexa
 
   class RequestHandler
 
-    attr_reader :request, :user_id
+    attr_reader :request, :user_id, :dialog_state
     def initialize(params)
       @request = AlexaRubykit.build_request params
       @user_id = User.find_or_create_by(amazon_user_id: request.session.user['userId']).id
+      @dialog_state = request.dialog_state
     end
 
     def process
-      handler = handler_for(request.type).new(self)
-      handler.process
+      if dialog_state.nil? || dialog_state == 'COMPLETED'
+        handler = handler_for(request.type).new(self)
+        handler.process
+      else
+        Helpers::HandlerHelper.create_delegate_response(intent: request.intent)
+      end
     end
 
     private
