+++ app/lib/handlers/intent_request_handler.rb
@@ -1,11 +1,12 @@
 module Handlers
 
-  class IntentRequestHandler < RequestHandler
+  class IntentRequestHandler
 
-    attr_reader :intent, :user_id
+    attr_reader :intent, :user_id, :dialog_state
     def initialize(context)
       @intent = context.request.intent
       @user_id = context.user_id
+      @dialog_state = context.dialog_state
     end
 
     attr_reader :handler
@@ -21,13 +22,13 @@ module Handlers
 
     def handler_for(intent)
       {
-        'Add'           => Handlers::TransactionIntentRequestHandler,
-        'Remove'        => Handlers::TransactionIntentRequestHandler,
-        'Warning'       => Handlers::ListIntentRequestHandler,
-        'Expired'       => Handlers::ListIntentRequestHandler,
-        'FromCategory'  => Handlers::ListIntentRequestHandler,
-        'AllItems'      => Handlers::ListIntentRequestHandler,
-        'GetExpiration' => Handlers::InformationIntentRequestHandler
+        'Add'           => Handlers::IntentRequestHandlers::TransactionIntentRequestHandler,
+        'Remove'        => Handlers::IntentRequestHandlers::TransactionIntentRequestHandler,
+        'Warning'       => Handlers::IntentRequestHandlers::ListIntentRequestHandler,
+        'Expired'       => Handlers::IntentRequestHandlers::ListIntentRequestHandler,
+        'FromCategory'  => Handlers::IntentRequestHandlers::ListIntentRequestHandler,
+        'AllItems'      => Handlers::IntentRequestHandlers::ListIntentRequestHandler,
+        'GetExpiration' => Handlers::IntentRequestHandlers::InformationIntentRequestHandler
       }[intent]
     end
