+++ app/lib/handlers/intent_request_handlers/base_request_handler.rb
@@ -1,0 +1,15 @@
+module Handlers
+  module IntentRequestHandlers
+
+    class BaseRequestHandler
+
+      attr_reader :intent, :user_id
+      def initialize(context)
+        @intent = context.intent
+        @user_id = context.user_id
+      end
+
+    end
+
+  end
+end
