--- app/lib/handlers/request_handler.rb
@@ -1,13 +1,0 @@
-module Handlers
-
-  class RequestHandler
-
-    attr_reader :intent, :user_id
-    def initialize(context)
-      @intent = context.intent
-      @user_id = context.user_id
-    end
-
-  end
-
-end
